requirements.txt
# Core LLM pipeline
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-deepseek>=0.1.0
langchain-mistralai>=0.2.0
langchain-together>=0.3.0
langchain-google-genai>=2.1.0
langsmith>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Structured output & label matching
rapidfuzz>=3.0.0
typing_extensions>=4.0.0

# ML / NLP
torch>=2.0.0
transformers>=4.40.0
datasets>=3.0.0
accelerate>=0.30.0
sentence-transformers>=3.0.0
sentencepiece>=0.2.0
safetensors>=0.4.0
huggingface-hub>=0.20.0
tiktoken>=0.5.0

# Data & analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.60.0

# API clients
openai>=1.0.0
requests>=2.28.0
